removing unused imports and fns pt1
