+++ zingolib/src/wallet/data.rs
@@ -1,306 +1,10 @@
 //! TODO: Add Mod Description Here!
 
-use byteorder::{LittleEndian, ReadBytesExt, WriteBytesExt};
-use incrementalmerkletree::witness::IncrementalWitness;
-use incrementalmerkletree::{Hashable, Position};
-
-use json::JsonValue;
-
-use std::convert::TryFrom;
 use std::io::{self, Read, Write};
 
-use zcash_encoding::{CompactSize, Optional};
+use byteorder::{LittleEndian, ReadBytesExt, WriteBytesExt};
 
-use zcash_primitives::memo::MemoBytes;
-use zcash_primitives::{memo::Memo, transaction::TxId};
-
-/// TODO: Add Doc Comment Here!
-pub const COMMITMENT_TREE_LEVELS: u8 = 32;
-/// TODO: Add Doc Comment Here!
-pub const MAX_SHARD_LEVEL: u8 = 16;
-
-/// TODO: Add Doc Comment Here!
-#[derive(Clone)]
-pub struct WitnessCache<Node: Hashable> {
-    /// TODO: Add Doc Comment Here!
-    pub(crate) witnesses: Vec<IncrementalWitness<Node, 32>>,
-    /// TODO: Add Doc Comment Here!
-    pub top_height: u64,
-}
-
-impl<Node: Hashable> std::fmt::Debug for WitnessCache<Node> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("WitnessCache")
-            .field("witnesses", &self.witnesses.len())
-            .field("top_height", &self.top_height)
-            .finish_non_exhaustive()
-    }
-}
-
-impl<Node: Hashable> WitnessCache<Node> {
-    /// TODO: Add Doc Comment Here!
-    pub fn new(witnesses: Vec<IncrementalWitness<Node, 32>>, top_height: u64) -> Self {
-        Self {
-            witnesses,
-            top_height,
-        }
-    }
-
-    /// TODO: Add Doc Comment Here!
-    pub fn empty() -> Self {
-        Self {
-            witnesses: vec![],
-            top_height: 0,
-        }
-    }
-
-    /// TODO: Add Doc Comment Here!
-    pub fn len(&self) -> usize {
-        self.witnesses.len()
-    }
-
-    /// TODO: Add Doc Comment Here!
-    pub fn is_empty(&self) -> bool {
-        self.witnesses.is_empty()
-    }
-
-    /// TODO: Add Doc Comment Here!
-    pub fn clear(&mut self) {
-        self.witnesses.clear();
-    }
-
-    /// TODO: Add Doc Comment Here!
-    pub fn get(&self, i: usize) -> Option<&IncrementalWitness<Node, 32>> {
-        self.witnesses.get(i)
-    }
-
-    /// TODO: Add Doc Comment Here!
-    #[cfg(test)]
-    pub fn get_from_last(&self, i: usize) -> Option<&IncrementalWitness<Node, 32>> {
-        self.witnesses.get(self.len() - i - 1)
-    }
-
-    /// TODO: Add Doc Comment Here!
-    pub fn last(&self) -> Option<&IncrementalWitness<Node, 32>> {
-        self.witnesses.last()
-    }
-
-    /// TODO: Add Doc Comment Here!
-    pub fn pop(&mut self, at_height: u64) {
-        while !self.witnesses.is_empty() && self.top_height >= at_height {
-            self.witnesses.pop();
-            self.top_height -= 1;
-        }
-    }
-
-    // pub fn get_as_string(&self, i: usize) -> String {
-    //     if i >= self.witnesses.len() {
-    //         return "".to_string();
-    //     }
-
-    //     let mut buf = vec![];
-    //     self.get(i).unwrap().write(&mut buf).unwrap();
-    //     return hex::encode(buf);
-    // }
-}
-
-// Reading a note also needs the corresponding address to read from.
-pub(crate) fn read_sapling_rseed<R: Read>(mut reader: R) -> io::Result<sapling_crypto::Rseed> {
-    let note_type = reader.read_u8()?;
-
-    let mut r_bytes: [u8; 32] = [0; 32];
-    reader.read_exact(&mut r_bytes)?;
-
-    let r = match note_type {
-        1 => sapling_crypto::Rseed::BeforeZip212(jubjub::Fr::from_bytes(&r_bytes).unwrap()),
-        2 => sapling_crypto::Rseed::AfterZip212(r_bytes),
-        _ => return Err(io::Error::new(io::ErrorKind::InvalidInput, "Bad note type")),
-    };
-
-    Ok(r)
-}
-
-pub(crate) fn write_sapling_rseed<W: Write>(
-    mut writer: W,
-    rseed: &sapling_crypto::Rseed,
-) -> io::Result<()> {
-    let note_type = match rseed {
-        sapling_crypto::Rseed::BeforeZip212(_) => 1,
-        sapling_crypto::Rseed::AfterZip212(_) => 2,
-    };
-    writer.write_u8(note_type)?;
-
-    match rseed {
-        sapling_crypto::Rseed::BeforeZip212(fr) => writer.write_all(&fr.to_bytes()),
-        sapling_crypto::Rseed::AfterZip212(b) => writer.write_all(b),
-    }
-}
-
-/// Only for TransactionRecords *from* "this" capability
-#[derive(Clone, Debug)]
-pub struct OutgoingTxData {
-    /// TODO: Add Doc Comment Here!
-    pub recipient_address: String,
-    /// Amount to this receiver
-    pub value: u64,
-    /// Note to the receiver, why not an option?
-    pub memo: Memo,
-    /// What if it wasn't provided?  How does this relate to
-    /// recipient_address?
-    pub recipient_ua: Option<String>,
-    /// This output's index in its containing transaction
-    pub output_index: Option<u64>,
-}
-impl std::fmt::Display for OutgoingTxData {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        // Format the recipient address or unified address if provided.
-        let address_display = if let Some(ref ua) = self.recipient_ua {
-            format!("unified address: {}", ua)
-        } else {
-            format!("recipient address: {}", self.recipient_address)
-        };
-        let memo_text = if let Memo::Text(mt) = self.memo.clone() {
-            mt.to_string()
-        } else {
-            "not a text memo".to_string()
-        };
-
-        write!(
-            f,
-            "\t{{
-            {}
-            value: {}
-            memo: {}
-        }}",
-            address_display, self.value, memo_text
-        )
-    }
-}
-impl From<OutgoingTxData> for JsonValue {
-    fn from(outgoing_tx_data: OutgoingTxData) -> Self {
-        let address = outgoing_tx_data
-            .recipient_ua
-            .unwrap_or(outgoing_tx_data.recipient_address);
-        let memo = if let Memo::Text(memo_text) = outgoing_tx_data.memo {
-            Some(memo_text.to_string())
-        } else {
-            None
-        };
-        json::object! {
-            "address" => address,
-            "value" => outgoing_tx_data.value,
-            "memo" => memo,
-        }
-    }
-}
-impl PartialEq for OutgoingTxData {
-    fn eq(&self, other: &Self) -> bool {
-        (self.recipient_address == other.recipient_address
-            || self.recipient_ua == other.recipient_ua)
-            && self.value == other.value
-            && self.memo == other.memo
-    }
-}
-impl OutgoingTxData {
-    const VERSION: u64 = 0;
-    /// Before version 0, OutgoingTxData didn't have a version field
-    pub fn read_old<R: Read>(mut reader: R) -> io::Result<Self> {
-        let address_len = reader.read_u64::<LittleEndian>()?;
-        let mut address_bytes = vec![0; address_len as usize];
-        reader.read_exact(&mut address_bytes)?;
-        let address = String::from_utf8(address_bytes).unwrap();
-
-        let value = reader.read_u64::<LittleEndian>()?;
-
-        let mut memo_bytes = [0u8; 512];
-        reader.read_exact(&mut memo_bytes)?;
-        let memo = match MemoBytes::from_bytes(&memo_bytes) {
-            Ok(mb) => match Memo::try_from(mb.clone()) {
-                Ok(m) => Ok(m),
-                Err(_) => Ok(Memo::Future(mb)),
-            },
-            Err(e) => Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                format!("Couldn't create memo: {}", e),
-            )),
-        }?;
-
-        Ok(OutgoingTxData {
-            recipient_address: address,
-            value,
-            memo,
-            recipient_ua: None,
-            output_index: None,
-        })
-    }
-
-    /// Read an OutgoingTxData from its serialized
-    /// representation
-    pub fn read<R: Read>(mut reader: R) -> io::Result<Self> {
-        let _external_version = CompactSize::read(&mut reader)?;
-        let address_len = reader.read_u64::<LittleEndian>()?;
-        let mut address_bytes = vec![0; address_len as usize];
-        reader.read_exact(&mut address_bytes)?;
-        let address = String::from_utf8(address_bytes).unwrap();
-
-        let value = reader.read_u64::<LittleEndian>()?;
-
-        let mut memo_bytes = [0u8; 512];
-        reader.read_exact(&mut memo_bytes)?;
-        let memo = match MemoBytes::from_bytes(&memo_bytes) {
-            Ok(mb) => match Memo::try_from(mb.clone()) {
-                Ok(m) => Ok(m),
-                Err(_) => Ok(Memo::Future(mb)),
-            },
-            Err(e) => Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                format!("Couldn't create memo: {}", e),
-            )),
-        }?;
-        let output_index = Optional::read(&mut reader, CompactSize::read)?;
-
-        Ok(OutgoingTxData {
-            recipient_address: address,
-            value,
-            memo,
-            recipient_ua: None,
-            output_index,
-        })
-    }
-
-    /// TODO: Add Doc Comment Here!
-    pub fn write<W: Write>(&self, mut writer: W) -> io::Result<()> {
-        CompactSize::write(&mut writer, Self::VERSION as usize)?;
-        // Strings are written as len + utf8
-        match &self.recipient_ua {
-            None => {
-                writer.write_u64::<LittleEndian>(self.recipient_address.as_bytes().len() as u64)?;
-                writer.write_all(self.recipient_address.as_bytes())?;
-            }
-            Some(ua) => {
-                writer.write_u64::<LittleEndian>(ua.as_bytes().len() as u64)?;
-                writer.write_all(ua.as_bytes())?;
-            }
-        }
-        writer.write_u64::<LittleEndian>(self.value)?;
-        writer.write_all(self.memo.encode().as_array())?;
-        Optional::write(&mut writer, self.output_index, |w, output_index| {
-            CompactSize::write(w, output_index as usize)
-        })
-    }
-}
-
-/// Wraps a vec of outgoing transaction data for the implementation of std::fmt::Display
-pub struct OutgoingTxDataSummaries(Vec<OutgoingTxData>);
-
-impl std::fmt::Display for OutgoingTxDataSummaries {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        for outgoing_tx_data in &self.0 {
-            write!(f, "\n{}", outgoing_tx_data)?;
-        }
-        Ok(())
-    }
-}
+use zcash_encoding::Optional;
 
 /// TODO: Add Mod Description Here!
 pub mod finsight {
@@ -1404,40 +1108,6 @@ pub mod summaries {
     }
 }
 
-/// Convenience wrapper for primitive
-#[derive(Debug)]
-pub struct SpendableSaplingNote {
-    /// TODO: Add Doc Comment Here!
-    pub transaction_id: TxId,
-    /// TODO: Add Doc Comment Here!
-    pub nullifier: sapling_crypto::Nullifier,
-    /// TODO: Add Doc Comment Here!
-    pub diversifier: sapling_crypto::Diversifier,
-    /// TODO: Add Doc Comment Here!
-    pub note: sapling_crypto::Note,
-    /// TODO: Add Doc Comment Here!
-    pub witnessed_position: Position,
-    /// TODO: Add Doc Comment Here!
-    pub extsk: Option<sapling_crypto::zip32::ExtendedSpendingKey>,
-}
-
-/// TODO: Add Doc Comment Here!
-#[derive(Debug)]
-pub struct SpendableOrchardNote {
-    /// TODO: Add Doc Comment Here!
-    pub transaction_id: TxId,
-    /// TODO: Add Doc Comment Here!
-    pub nullifier: orchard::note::Nullifier,
-    /// TODO: Add Doc Comment Here!
-    pub diversifier: orchard::keys::Diversifier,
-    /// TODO: Add Doc Comment Here!
-    pub note: orchard::note::Note,
-    /// TODO: Add Doc Comment Here!
-    pub witnessed_position: Position,
-    /// TODO: Add Doc Comment Here!
-    pub spend_key: Option<orchard::keys::SpendingKey>,
-}
-
 /// Struct that tracks the latest and historical price of ZEC in the wallet
 #[derive(Clone, Debug)]
 pub struct WalletZecPriceInfo {
@@ -1512,65 +1182,5 @@ impl WalletZecPriceInfo {
         writer.write_u64::<LittleEndian>(self.historical_prices_retry_count)?;
 
         Ok(())
-    }
-}
-
-#[cfg(any(test, feature = "test-elevation"))]
-pub(crate) mod mocks {
-    use zcash_primitives::memo::Memo;
-
-    use crate::utils::build_method;
-
-    use super::OutgoingTxData;
-
-    pub(crate) struct OutgoingTxDataBuilder {
-        recipient_address: Option<String>,
-        value: Option<u64>,
-        memo: Option<Memo>,
-        recipient_ua: Option<Option<String>>,
-        output_index: Option<Option<u64>>,
-    }
-
-    impl OutgoingTxDataBuilder {
-        pub(crate) fn new() -> Self {
-            Self {
-                recipient_address: None,
-                value: None,
-                memo: None,
-                recipient_ua: None,
-                output_index: None,
-            }
-        }
-
-        // Methods to set each field
-        build_method!(recipient_address, String);
-        build_method!(value, u64);
-        build_method!(memo, Memo);
-        build_method!(recipient_ua, Option<String>);
-        build_method!(output_index, Option<u64>);
-
-        pub(crate) fn build(&self) -> OutgoingTxData {
-            OutgoingTxData {
-                recipient_address: self.recipient_address.clone().unwrap(),
-                value: self.value.unwrap(),
-                memo: self.memo.clone().unwrap(),
-                recipient_ua: self.recipient_ua.clone().unwrap(),
-                output_index: self.output_index.unwrap(),
-            }
-        }
-    }
-
-    impl Default for OutgoingTxDataBuilder {
-        fn default() -> Self {
-            let mut builder = Self::new();
-            builder
-                .recipient_address("default_address".to_string())
-                .value(50_000)
-                .memo(Memo::default())
-                .recipient_ua(None)
-                .output_index(None);
-
-            builder
-        }
     }
 }
